Prepare for an 0.6.0 release of the Viam Triton Module
- Upgrade Jetson from r35.4 to r35.5
- Upgrade Viam C++ SDK to v0.7.0
- Downgrade Triton from 2.35 to 2.34

The Triton downgrade is unfortunate, but we are left with little
choice given that NVidia has pulled the only working builds of
2.35. Fortunately, we do not appear to depend on any features added
in 2.35.

See https://github.com/triton-inference-server/server/issues/6844 for
the frustrating details.

This change also includes substantial changes to the docker build made
in an effort to improve layer caching during builds.