Use tags when deploying to the module registry (#17)
We now have just a single module, but 3 sets of tags for it. 

Alas, the way to check if this works is to merge the PR and then make a pre-release version of the module, so there isn't an easy way to try it out pre-merging.

Abe, does this look right to you? If so, and if the pre-release builds look right, I should manually upload the release versions of the module, so they're in the right order until tag priority is a thing.


* add tags to GH Action, need to simplify meta.json again

* feedback from Abe

* switch back to a single module. Don't put the words CUDA or Jetpack in the name, since it does both

* all builds go to the same module