use the original module name: making a new one has model name conflicts (#18)
One of these days I'll get it right.

With the old name, we had errors publishing to the registry because there were model name conflicts between two modules (the original module name and the new one that didn't mention Jetpack). Ideally, we'll rename this module to not mention Jetpack, since you can now run it without it.