More make targets (#27)
The targets used in the CI system should be unchanged, but these make it easier to run manually, too. I've done pretty extensive tests of the `module-` targets (and by implication the `.sh` ones), but only cursory tests of the `image-` ones. but the latter all start running docker builds with believable-looking tags!


* add make targets for specific containers, probably not debugged yet

* add targets for different modules

* simplify to remove repetition

* don't hide comment after envsubst

* do the same boiler-plate removal in the docker images

* more exact terminology: we're making images, not containers

* tweak comment

* rejigger to cut down even more on duplication, and bring back the part that the CI system uses

* make anything that depends on the tag .PHONY

* sigh. duplicate rule because .PHONY doesn't work with patterns

* more whitespace: 2 blank lines between major sections

* remove unnecessary .PHONY rule

* docker tags must be all-lowercase