[RDSK-9415] remove parallelism when uploading to the registry (#32)
With this change, deploying to Viam's modular registry is done serially, so they should be in the order listed (jetpack 5, then jetpack 6, and finally cuda). This should be the right order: without tag priorities (which are in a [backlogged ticket](https://viam.atlassian.net/browse/APP-6759), your machine will get the earliest-uploaded module with compatible tags (so if your machine has both jetpack and cuda support, you'll get the jetpack version).

Thanks to Abe for talking through this with me!

I've also made some minor changes to bring this inline with Bijan's checklist for standardizing a CV module (renaming the file, having it trigger on the tags instead of the release type). 

I'm unsure how to actually test this out before merging it, though. Any ideas?


* remove parallelism when uploading to the registry

* use the tag format from Bijan's checklist

* rename deploy.yml -> publish.yml in keeping with Bijan's checklist

* add comments summarizing what each job does