Workflow tweak (#33)
Thanks to Drew for suggesting we change the CV checklist, and thanks to Bijan for looking into how to validate that the tag looks right.

Compare this to https://github.com/viamrobotics/ocean-prefilter/blob/main/.github/workflows/build-arm64-module.yml, which Bijan says works correctly.

I have not tried this out: it's unclear to me how to do that before merging.


* add step to validate semver tag when publishing a release

* get all steps to depend on validating the tag