Remove input tensor memory type and device
It isn't clear that these settings were ever appropriate or meaningful
in a single GPU environment, and there is no way to use them correctly
in a multi-GPU environment since it is impossible to know which GPU
the Triton infrastructure is going to route the request to. It might
be useful again in the future, if we had a way for RDK to allocate
input tensors (or obtain output tensors) via shared memory regions
that could be "on GPU", but as of now we have no ability to do that.